requirements.txt
Django==5.0.7
python-dotenv==1.0.1
psycopg2-binary==2.9.6
gunicorn==20.1.0
django-crispy-forms==1.14.0
crispy-bootstrap5
django-filter>=23.5
django-auditlog==2.3.0
whitenoise==6.5.0